Adds direct whisper capability Adds Base class for Layouting: New Exalt and Divine Formatters
