+++ frontend/getting-rich/src/app/tradingview/whispers/whispers.component.html
@@ -1,7 +1,7 @@
 <h2>Available Whispers</h2>
-<button mat-button (keydown.control.c)="onKeyDown()" (mouseup)="copyNewWhisper()" [cdkCopyToClipboard]="newest_whisper">Copy new whisper</button>
+<button mat-button (keydown.control.c)="onKeyDown()" (mouseup)="copyNewWhisper()">Copy new whisper</button>
 <div>
-  <div *ngFor="let whisper of newWhispers" style="justify-content: space-between" matTooltip="{{whisper.whisper}}">
+  <div *ngFor="let whisper of newWhispers" style="justify-content: space-between" matTooltip="{{whisper.whisper}}" (click)="directWhisper(whisper.whisper_token,whisper.offer_count)">
     <span class="profit">+{{whisper.profit | number : '.2-2'}} {{(whisper.value-whisper.chaos_price)/whisper.chaos_price |percent}}</span>
     <span class="whisper" >{{whisper.whisper  | slice:0:60 }}...</span>
   </div>
