+++ frontend/thinking/trades/views.py
@@ -1,18 +1,21 @@
 import json
 from pathlib import Path
 
+import requests
 from django.http import JsonResponse
 
 # Create your views here.
 from constants import LEAGUE
 from poe.ninja import retrieve_prices
+from poe.trade.exchange_parser import exchange_parser
 from poe.trade.exchange_resolver import ExchangeResolver
+from poe.trade.headers import headers
 from poe.trade.listings_resolver import ListingsResolver
 import pandas as pd
 
 from poe.trade.search_resolver import SearchResolver
 from poe.trade_finder import find_profitable_items
-from poe.trade_finder.whisper_generator import WhisperGenerator, exchange_parser
+from poe.trade_finder.whisper_generator.copy_whisper import CopyWhisperGenerator
 from poe.valuation import own_valuations
 from poe.valuation.gems.experience import xp_value
 from poe.valuation.gems.regrading_lens import (
@@ -44,21 +47,11 @@ def profitable_items(request):
 
 
 def whispers(request):
-    items = (
-        pd.DataFrame(list(zip(*dict(request.GET).values())), columns=request.GET.keys())
-        .rename({"name": "index"}, axis=1)
-        .set_index("index")
-    )
-    items["expected_profit"] = items["expected_profit"].astype("float")
-    items["value"] = items["value"].astype("float")
-    prices = {item.name: [{"chaosValue": item.price}] for item in Item.objects.all()}
-    key_mapping = pd.read_csv(
-        f"{Path(__file__).resolve().parent}/poe_keys.csv"
-    ).set_index("name")["key"]
+    items, key_mapping, prices = mymethod(request)
 
     listings_resolver = ListingsResolver(league=LEAGUE)
     exchange_resolver = ExchangeResolver(league=LEAGUE)
-    use_case = WhisperGenerator(
+    use_case = CopyWhisperGenerator(
         exchange_resolver=exchange_resolver,
         listings_resolver=listings_resolver,
         poe_trade_key_mapping=key_mapping,
@@ -67,6 +60,21 @@ def whispers(request):
     domain_result = use_case(items)
     result = list(domain_result.T.to_dict().values())
     return JsonResponse(result, safe=False)
+
+
+def mymethod(request):
+    items = (
+        pd.DataFrame(list(zip(*dict(request.GET).values())), columns=request.GET.keys())
+            .rename({"name": "index"}, axis=1)
+            .set_index("index")
+    )
+    items["expected_profit"] = items["expected_profit"].astype("float")
+    items["value"] = items["value"].astype("float")
+    prices = {item.name: [{"chaosValue": item.price}] for item in Item.objects.all()}
+    key_mapping = pd.read_csv(
+        f"{Path(__file__).resolve().parent}/poe_keys.csv"
+    ).set_index("name")["key"]
+    return items, key_mapping, prices
 
 
 def orb_of_horizons(request):
@@ -191,3 +199,11 @@ def regrading_lens(request):
 #     return JsonResponse(result, safe=False)
 def fragment_sets():
     pass
+def direct_whisper(request):
+    params = json.loads(request.body)
+    url= 'https://www.pathofexile.com/api/trade/whisper'
+
+    response=requests.post(url,json=params,headers={    'x-requested-with': 'XMLHttpRequest', **headers})
+    response.raise_for_status()
+    return response.json()
+
