+++ poe/sale/inventory_management_builder.py
@@ -2,7 +2,8 @@ from fractions import Fraction
 
 from poe.item.item_builder import ItemBuilder
 from poe.sale.inventory_creator import InventoryCreator
-from poe.sale.layouters.base import Layouter
+from poe.sale.layouters.divine import DivineLayouter
+from poe.sale.layouters.exalt import ExaltLayouter
 from poe.sale.seller import Seller
 from poe.sale.up_pricer import Pricer
 from poe.sale.inventory_management import InventoryManagement
@@ -21,9 +22,9 @@ def build_inventory_management(
 ):
     item_builder = ItemBuilder(prices, valuations)
     inventory_creator = InventoryCreator()
-    exalt_value = prices["Exalted Orb"][0]["chaosValue"]
+    divine_value = prices["Divine Orb"][0]["chaosValue"]
     pricer = Pricer(price_base_by_type, price_base)
-    layouter = Layouter(exalt_value=exalt_value, min_trade_value=min_trade_value)
+    layouter = DivineLayouter(divine_value=divine_value, min_trade_value=min_trade_value)
     seller = Seller(
         up_price=pricer.up_price, cleaning_rules=cleaning_rules, layouter=layouter
     )
