+++ poe/sale/seller.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from typing import Callable
 
 from poe.sale.layouters.base import Layouter
+from poe.sale.layouters.exalt import ExaltLayouter
 
 
 @dataclass
