+++ poe/trade/headers.py
@@ -2,6 +2,10 @@ from constants import ssid
 
 headers = {
     "authority": "www.pathofexile.com",
+    'accept': '*/*',
+    'accept-language': 'de-DE,de;q=0.9,en-US;q=0.8,en;q=0.7,es;q=0.6,pl;q=0.5,sl;q=0.4',
     "user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/89.0.4389.82 Safari/537.36",
-    "cookie": f"__cfduid=d18f7bd987ce2b3f98c02b05b92a782c31618165141; cf_clearance=7d322aa725fafce9ef1ebc8a26759033b127dd7f-1618777267-0-150; POESESSID={ssid}; _ga=GA1.2.1675387988.1618777334; _gid=GA1.2.715037690.1618777334",
+    'cookie': f'cf_clearance=jwZmq02SAhIZjXQQRU4oWAdyF2pA2lR0gz7vpMFxOt0-1654637582-0-150; POESESSID={ssid}',
+    'dnt': '1',
+    'origin': 'https://www.pathofexile.com',
 }
